(PC-32146)[API] refactor: rename ubble error file
models.py files are reserved for database models
